Add a flake.nix to the repository
This does allow a full build with `nix build` and get all dependencies
for local development in a shell with `nix develop`. From within that
shell normal invocations of `cargo` will just work.

This also adds a .envrc with which direnv users can allow a nix shell to
be loaded and cached automatically.